+++ package.json
@@ -4,7 +4,7 @@
   "description": "A platform for rating online courses",
   "main": "index.js",
   "scripts": {
-    "test": "node tests",
+    "test": "mocha",
     "lint": "./node_modules/.bin/eslint .",
     "lint:fix": "./node_modules/.bin/eslint . --fix",
     "start": "node start.js",
@@ -18,6 +18,7 @@
     "db-migrate-pg": "^0.5.0",
     "dotenv": "^7.0.0",
     "express": "^4.16.4",
+    "mocha": "^6.0.2",
     "morgan": "^1.9.1",
     "pg": "^7.9.0",
     "pg-hstore": "^2.3.2",
